Extract traits and case classes into common
