+++ src/common/src/main/scala/osmesa/common/model/impl/package.scala
@@ -1,4 +1,4 @@
-package osmesa.analytics.oneoffs
+package osmesa.common.model
 
 import geotrellis.raster.{
   CellType,
@@ -8,63 +8,9 @@ import geotrellis.raster.{
   Tile,
   Raster => GTRaster
 }
-import geotrellis.vector.io._
-import geotrellis.vector.{Extent, Point, Geometry => GTGeometry}
+import geotrellis.vector.Extent
 
-// this is a placeholder package until such point at which much of this is moved to core (with case classes in an impl
-// package)
-package object traits {
-  trait Geometry {
-    def geom: GTGeometry
-  }
-
-  trait SerializedGeometry extends Geometry {
-    lazy val geom: GTGeometry = wkb.readWKB
-
-    def wkb: Array[Byte]
-  }
-
-  trait TileCoordinates {
-    def zoom: Int
-    def x: Int
-    def y: Int
-  }
-
-  trait GeometryTile extends SerializedGeometry with TileCoordinates
-
-  trait Raster {
-    def raster: GTRaster[Tile]
-  }
-
-  trait RasterTile extends Raster with TileCoordinates
-
-  trait Coordinates extends Geometry {
-    def lat: Option[BigDecimal]
-    def lon: Option[BigDecimal]
-
-    def geom: Point = Point(x, y)
-
-    def x: Float = lon.map(_.floatValue).getOrElse(Float.NaN)
-    def y: Float = lat.map(_.floatValue).getOrElse(Float.NaN)
-  }
-
-  trait Key {
-    def key: String
-  }
-
-  trait Sequence {
-    def sequence: Int
-  }
-
-  // NOTE this doesn't extend TileSeq[T] to avoid using type parameters
-  trait RasterWithSequenceTileSeq {
-    def tiles: Seq[Raster with Sequence]
-  }
-
-  trait Count {
-    def count: Long
-  }
-
+package object impl {
   case class CoordinatesWithKeyAndSequence(sequence: Int,
                                            key: String,
                                            lat: Option[BigDecimal],
